requirements.prod.txt
# Production runtime dependencies only
# (Training, notebooks, linting tools are intentionally excluded)

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Database (Postgres + pgvector via SQLAlchemy)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg[binary]>=3.1.0
pgvector>=0.2.5

# ML runtime
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# LLM (optional — graceful fallback if ANTHROPIC_API_KEY not set)
anthropic>=0.40.0
